requirements.txt
flask
requests
google-genai
sift-stack-py
google-generativeai